+++ src/main/java/ladysnake/satin/impl/RenderLayerSupplier.java
@@ -21,6 +21,8 @@ import ladysnake.satin.mixin.client.render.RenderPhaseAccessor;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.RenderPhase;
 import net.minecraft.client.render.Shader;
+import net.minecraft.client.render.VertexFormat;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -31,6 +33,7 @@ public class RenderLayerSupplier {
     private final Consumer<RenderLayer.MultiPhaseParameters.Builder> transform;
     private final Map<RenderLayer, RenderLayer> renderLayerCache = new HashMap<>();
     private final String uniqueName;
+    private final @Nullable VertexFormat vertexFormat;
 
     public static RenderLayerSupplier framebuffer(String name, Runnable setupState, Runnable cleanupState) {
         RenderPhase.Target target = new RenderPhase.Target(
@@ -41,13 +44,18 @@ public class RenderLayerSupplier {
         return new RenderLayerSupplier(name, builder -> builder.target(target));
     }
 
-    public static RenderLayerSupplier shader(String name, Supplier<Shader> shaderSupplier) {
+    public static RenderLayerSupplier shader(String name, VertexFormat vertexFormat, Supplier<Shader> shaderSupplier) {
         RenderPhase shader = Helper.makeShader(shaderSupplier);
-        return new RenderLayerSupplier(name, builder -> Helper.applyShader(builder, shader));
+        return new RenderLayerSupplier(name, vertexFormat, builder -> Helper.applyShader(builder, shader));
     }
 
-    public RenderLayerSupplier(String name, Consumer<RenderLayer.MultiPhaseParameters.Builder> transformer) {
+    public RenderLayerSupplier(String name,  Consumer<RenderLayer.MultiPhaseParameters.Builder> transformer) {
+        this(name, null, transformer);
+    }
+
+    public RenderLayerSupplier(String name, @Nullable VertexFormat vertexFormat, Consumer<RenderLayer.MultiPhaseParameters.Builder> transformer) {
         this.uniqueName = name;
+        this.vertexFormat = vertexFormat;
         this.transform = transformer;
     }
 
@@ -57,7 +65,7 @@ public class RenderLayerSupplier {
             return existing;
         }
         String newName = ((RenderPhaseAccessor) baseLayer).getName() + "_" + this.uniqueName;
-        RenderLayer newLayer = RenderLayerDuplicator.copy(baseLayer, newName, this.transform);
+        RenderLayer newLayer = RenderLayerDuplicator.copy(baseLayer, newName, this.vertexFormat, this.transform);
         this.renderLayerCache.put(baseLayer, newLayer);
         return newLayer;
     }